+++ src/utils/AppConfig.ts
@@ -4,15 +4,13 @@ import { BILLING_INTERVAL, type PricingPlan } from '@/types/Subscription';
 
 const localePrefix: LocalePrefix = 'as-needed';
 
-// FIXME: Update this configuration file based on your project information
 export const AppConfig = {
-  name: 'SaaS Template',
+  name: 'Cutting Edge Chat',
   locales: [
     {
       id: 'en',
       name: 'English',
     },
-    { id: 'fr', name: 'Français' },
   ],
   defaultLocale: 'en',
   localePrefix,
@@ -35,20 +33,19 @@ export const PricingPlanList: Record<string, PricingPlan> = {
     devPriceId: '',
     prodPriceId: '',
     features: {
-      teamMember: 2,
-      website: 2,
-      storage: 2,
-      transfer: 2,
+      teamMember: 1,
+      website: 1,
+      storage: 1,
+      transfer: 1,
     },
   },
   [PLAN_ID.PREMIUM]: {
     id: PLAN_ID.PREMIUM,
     price: 79,
     interval: BILLING_INTERVAL.MONTH,
-    testPriceId: 'price_premium_test', // Use for testing
-    // FIXME: Update the price ID, you can create it after running `npm run stripe:setup-price`
-    devPriceId: 'price_1PNksvKOp3DEwzQlGOXO7YBK',
-    prodPriceId: '',
+    testPriceId: 'price_premium_test',
+    devPriceId: 'price_1TTJp432TnnnYrTabzx3UdX9',
+    prodPriceId: 'price_1TTJp432TnnnYrTabzx3UdX9',
     features: {
       teamMember: 5,
       website: 5,
@@ -60,10 +57,9 @@ export const PricingPlanList: Record<string, PricingPlan> = {
     id: PLAN_ID.ENTERPRISE,
     price: 199,
     interval: BILLING_INTERVAL.MONTH,
-    testPriceId: 'price_enterprise_test', // Use for testing
-    // FIXME: Update the price ID, you can create it after running `npm run stripe:setup-price`
-    devPriceId: 'price_1PNksvKOp3DEwzQli9IvXzgb',
-    prodPriceId: 'price_123',
+    testPriceId: 'price_enterprise_test',
+    devPriceId: 'price_1TTJp432TnnnYrTabzx3UdX9',
+    prodPriceId: 'price_1TTJp432TnnnYrTabzx3UdX9',
     features: {
       teamMember: 100,
       website: 100,